+++ tests/test_arrow.py
@@ -1935,8 +1935,28 @@ class TestArrowHumanize:
         assert later506.humanize(self.now, granularity="week") == "in 82 weeks"
         assert self.now.humanize(later506, granularity="month") == "18 months ago"
         assert later506.humanize(self.now, granularity="month") == "in 18 months"
+        assert self.now.humanize(later506, granularity="quarter") == "6 quarters ago"
+        assert later506.humanize(self.now, granularity="quarter") == "in 6 quarters"
         assert self.now.humanize(later506, granularity="year") == "a year ago"
         assert later506.humanize(self.now, granularity="year") == "in a year"
+
+        assert self.now.humanize(later1, granularity="quarter") == "0 quarters ago"
+        assert later1.humanize(self.now, granularity="quarter") == "in 0 quarters"
+        later107 = self.now.shift(seconds=10 ** 7)
+        assert self.now.humanize(later107, granularity="quarter") == "a quarter ago"
+        assert later107.humanize(self.now, granularity="quarter") == "in a quarter"
+        later207 = self.now.shift(seconds=2 * 10 ** 7)
+        assert self.now.humanize(later207, granularity="quarter") == "2 quarters ago"
+        assert later207.humanize(self.now, granularity="quarter") == "in 2 quarters"
+        later307 = self.now.shift(seconds=3 * 10 ** 7)
+        assert self.now.humanize(later307, granularity="quarter") == "3 quarters ago"
+        assert later307.humanize(self.now, granularity="quarter") == "in 3 quarters"
+        later377 = self.now.shift(seconds=3.7 * 10 ** 7)
+        assert self.now.humanize(later377, granularity="quarter") == "4 quarters ago"
+        assert later377.humanize(self.now, granularity="quarter") == "in 4 quarters"
+        later407 = self.now.shift(seconds=4 * 10 ** 7)
+        assert self.now.humanize(later407, granularity="quarter") == "5 quarters ago"
+        assert later407.humanize(self.now, granularity="quarter") == "in 5 quarters"
 
         later108 = self.now.shift(seconds=10 ** 8)
         assert self.now.humanize(later108, granularity="year") == "3 years ago"
@@ -2309,6 +2329,13 @@ class TestArrowHumanize:
         with pytest.raises(ValueError):
             arw.humanize(later, granularity="week")
 
+    def test_empty_granularity_list(self):
+        arw = arrow.Arrow(2013, 1, 1, 0, 0, 0)
+        later = arw.shift(seconds=55000)
+
+        with pytest.raises(ValueError):
+            arw.humanize(later, granularity=[])
+
     # Bulgarian is an example of a language that overrides _format_timeframe
     # Applicabale to all locales. Note: Contributors need to make sure
     # that if they override describe or describe_mutli, that delta
@@ -2343,8 +2370,7 @@ class TestArrowHumanizeTestsWithLocale:
         arw = arrow.Arrow(2013, 1, 1, 0, 0, 44)
 
         result = arw.humanize(self.datetime, locale="ru")
-
-        assert result == "через 44 несколько секунд"
+        assert result == "через 44 секунды"
 
     def test_years(self):